멀리 뛰기 수정

+++ challenge_Java/멀리 뛰기.java
@@ -2,7 +2,7 @@ class Solution {
     public long solution(int n) {
         if (n == 1) return 1;
         
-        long[] jump = new long[n];
+        int[] jump = new int[n];
         jump[0] = 1;
         jump[1] = 2;
         for (int i = 2; i < n; i++)
